Add modal-reducer
